move styles to seperate package
