+++ packages/jarveybot-react/src/ui/ChatHeader.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import CloseIcon from '../icons/Close';
-import './ChatHeader.scss';
+import '../../../jarveybot-styles/src/Header.scss';
 
 interface PropsInterface {
   handleCloseChat?: Function;
