+++ packages/jarveybot-styles/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@jarveybot/styles",
+  "description": "A chat bot library, which you can use as react component, react hook, or vanilla javascript module.",
+  "keywords": [
+    "Chatbot",
+    "React Component",
+    "React Hook",
+    "Vue Component",
+    "Vue Composiotion function"
+  ],
+  "version": "0.1.0",
+  "author": "Sebastian Sauer <sebsauer90@gmail.com>",
+  "license": "MIT",
+  "private": false,
+  "devDependencies": {},
+  "scripts": {}
+}
