+++ packages/jarveybot-styles/src/Chat.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "variables";
 
 .jb {
   &__container {
