+++ packages/jarveybot-styles/src/variables.scss
